Couple other cleanups on MPA
